Change header stats ui
